requirements.txt
google-generativeai
python-dotenv
pandas
pypdf
numpy